fix(billing): co-locate BYOT provider + key with workspace billing
Reverts the /admin/model-config → /platform/model-config move from #2305
and co-locates BYOT setup with the BYOT toggle on /admin/billing so the
user no longer bounces between pages.

The audit agent that drove #2305's classification pattern-matched on
the page's `usePlatformAdminGuard` import and concluded the page was
platform-tier. It missed that the API at /api/v1/admin/model-config
uses `createAdminRouter()` + `requirePermission("admin:settings")`
and scopes everything by the active org. BYOT is per-workspace — every
org configures their own provider/key, exactly like /admin/billing —
and the page-level guard was the actual bug.

Effect of the regression: workspace admins flipping the BYOT switch on
/admin/billing hit a CTA pointing at /platform/model-config, where the
platform guard `router.replace("/admin")` redirected them away from the
key form. With no other path, BYOT was effectively unreachable on SaaS
for non-platform admins.

Shape

- New `<ModelProviderSection mode="page" | "billing" />` lives at
  packages/web/src/ui/components/admin/model-provider-section.tsx and
  owns the form state, catalog fetches, save/delete/test mutations,
  and the BYOT-eligibility gate.
- /admin/model-config is a thin wrapper: hero + platform-baseline row
  + `<ModelProviderSection mode="page" />`. usePlatformAdminGuard is
  gone.
- /admin/billing's `ByotRow` mounts `<ModelProviderSection mode="billing" />`
  inline when the toggle is on. The "Add your API key" / "Manage"
  buttons that used to hand the user off to a separate page are gone
  — the form is right there.
- admin-nav.ts: /admin/model-config returns to the Intelligence group
  (no requiredRole — workspace admins should see it); the Platform
  entry that #2305 added is dropped.

API paths are untouched. /api/v1/admin/model-config/* stays at its
current mount, which has always been org-scoped.

Don't touch

- /platform/plugin-registry is still platform-tier (uses
  createPlatformRouter()) — that move from #2305 was correct.
- The /platform/users split from #2307 is a real workspace-vs-platform
  split where both pages exist intentionally.
- ENROLLMENT_URL in admin-mfa-required.ts continues to point at
  /admin/account-security.

Tests

- byot-key-status.test.tsx deleted — the component it guarded (a
  presenter for the cross-page CTA) no longer exists; the regression
  it guarded against (#2172: no path to add a key) is now
  structurally prevented by the inline form mount.
- admin-nav.test.ts's "every nav href has a matching page.tsx"
  walker covers /admin/model-config; comment updated to record the
  partial revert.
- Full /ci sweep clean: lint, type, test (passkey-tile flake passed
  in isolation), syncpack, template drift, security headers,
  railway-watch, schema drift, oauth-helper drift.
